+++ website/docusaurus.config.js
@@ -24,7 +24,8 @@ const config = {
         docs: {
           sidebarPath: require.resolve('./sidebars.js'),
           // Please change this to your repo.
-          editUrl: 'https://github.com/XHMM/react-native-cloud-store/edit/main/website/',
+          editUrl:
+            'https://github.com/XHMM/react-native-cloud-store/edit/main/website/',
 
           // uncomment when developing